+++ bbot/test/test_step_2/module_tests/test_module_bucket_amazon.py
@@ -26,11 +26,15 @@ class Bucket_Amazon_Base(ModuleTestBase):
 
     @property
     def config_overrides(self):
-        return {"modules": {f"bucket_{self.provider}": {"permutations": True}}}
+        return {"modules": {self.module_name: {"permutations": True}}}
+
+    @property
+    def module_name(self):
+        return self.__class__.__name__.lower().split("test")[-1]
 
     @property
     def modules_overrides(self):
-        return ["excavate", "speculate", "httpx", f"bucket_{self.provider}"]
+        return ["excavate", "speculate", "httpx", self.module_name]
 
     def url_setup(self):
         self.url_1 = f"https://{self.random_bucket_1}"
@@ -68,14 +72,14 @@ class Bucket_Amazon_Base(ModuleTestBase):
         # make sure buckets were excavated
         assert any(
             e.type == "STORAGE_BUCKET" and str(e.module) == f"{self.provider}_cloud" for e in events
-        ), f'bucket not found for provider "{self.provider}"'
+        ), f'bucket not found for module "{self.module_name}"'
         # make sure open buckets were found
         if module_test.module.supports_open_check:
             assert any(
-                e.type == "FINDING" and str(e.module) == f"bucket_{self.provider}" for e in events
-            ), f'open bucket not found for provider "{self.provider}"'
+                e.type == "FINDING" and str(e.module) == self.module_name for e in events
+            ), f'open bucket not found for module "{self.module_name}"'
             for e in events:
-                if e.type == "FINDING" and str(e.module) == f"bucket_{self.provider}":
+                if e.type == "FINDING" and str(e.module) == self.module_name:
                     url = e.data.get("url", "")
                     assert self.random_bucket_2 in url
                     assert not self.random_bucket_1 in url
@@ -83,10 +87,10 @@ class Bucket_Amazon_Base(ModuleTestBase):
         # make sure bucket mutations were found
         assert any(
             e.type == "STORAGE_BUCKET"
-            and str(e.module) == f"bucket_{self.provider}"
+            and str(e.module) == self.module_name
             and f"{random_bucket_name_3}" in e.data["url"]
             for e in events
-        ), f'bucket (dev mutation) not found for provider "{self.provider}"'
+        ), f'bucket (dev mutation) not found for module "{self.module_name}"'
 
 
 class TestBucket_Amazon(Bucket_Amazon_Base):
