+++ bbot/test/test_step_2/module_tests/test_module_bucket_firebase.py
@@ -2,7 +2,7 @@ from .test_module_bucket_amazon import *
 
 
 class TestBucket_Firebase(Bucket_Amazon_Base):
-    provider = "firebase"
+    provider = "google"
     random_bucket_1 = f"{random_bucket_name_1}.firebaseio.com"
     random_bucket_2 = f"{random_bucket_name_2}.firebaseio.com"
     random_bucket_3 = f"{random_bucket_name_3}.firebaseio.com"
